+++ app/src/main/java/com/concordium/wallet/ui/more/moreoverview/MoreOverviewFragment.kt
@@ -117,7 +117,10 @@ class MoreOverviewFragment : BaseFragment() {
         )
 
         binding.identitiesLayout.setOnClickListener {
-            gotoIdentitiesOverview()
+            if (mainViewModel.hasCompletedOnboarding())
+                gotoIdentitiesOverview()
+            else
+                (requireActivity() as BaseActivity).showUnlockFeatureDialog()
         }
 
         binding.walletsLayout.setOnClickListener {
@@ -133,7 +136,10 @@ class MoreOverviewFragment : BaseFragment() {
         }
 
         binding.exportLayout.setOnClickListener {
-            gotoExport()
+            if (mainViewModel.hasCompletedOnboarding())
+                gotoExport()
+            else
+                (requireActivity() as BaseActivity).showUnlockFeatureDialog()
         }
         viewModel.fileImportExportVisibilityLiveData.observe(
             viewLifecycleOwner,
@@ -157,7 +163,10 @@ class MoreOverviewFragment : BaseFragment() {
         )
 
         binding.seedPhraseRevealLayout.setOnClickListener {
-            revealSeedPhrase()
+            if (mainViewModel.hasCompletedOnboarding())
+                revealSeedPhrase()
+            else
+                (requireActivity() as BaseActivity).showUnlockFeatureDialog()
         }
         viewModel.seedPhraseRevealVisibilityLiveData.observe(
             viewLifecycleOwner,
@@ -165,7 +174,10 @@ class MoreOverviewFragment : BaseFragment() {
         )
 
         binding.seedRevealLayout.setOnClickListener {
-            revealSeed()
+            if (mainViewModel.hasCompletedOnboarding())
+                revealSeed()
+            else
+                (requireActivity() as BaseActivity).showUnlockFeatureDialog()
         }
         viewModel.seedRevealVisibilityLiveData.observe(
             viewLifecycleOwner,