+++ app/src/main/java/com/concordium/wallet/ui/onboarding/OnboardingState.kt
@@ -1,0 +1,10 @@
+package com.concordium.wallet.ui.onboarding
+
+enum class OnboardingState {
+    SAVE_PHRASE,
+    VERIFY_IDENTITY,
+    IDENTITY_IN_PROGRESS,
+    IDENTITY_UNSUCCESSFUL,
+    CREATE_ACCOUNT,
+    DONE
+}