+++ app/src/main/java/com/concordium/wallet/ui/seed/recover/seed/SeedRecoverViewModel.kt
@@ -59,7 +59,9 @@ class SeedRecoverViewModel(application: Application) : AndroidViewModel(applicat
                 newWalletType = AppWallet.Type.SEED,
             )
             App.appCore.setup.finishInitialSetup()
+            App.appCore.session.walletStorage.setupPreferences.setHasCompletedOnboarding(true)
         }
+
         _saveSeedSuccess.value = success
     }