+++ app/src/main/java/com/concordium/wallet/ui/welcome/WelcomeViewModel.kt
@@ -1,0 +1,11 @@
+package com.concordium.wallet.ui.welcome
+
+import android.app.Application
+import androidx.lifecycle.AndroidViewModel
+import com.concordium.wallet.App
+
+class WelcomeViewModel(application: Application) : AndroidViewModel(application) {
+
+    val shouldSetUpPassword: Boolean
+        get() = !App.appCore.setup.isAuthSetupCompleted
+}